refactor(module:divider): refactor divider (#4780)
