--- components/divider/nz-divider.component.html
@@ -1,3 +1,0 @@
-<span *ngIf="nzText" class="ant-divider-inner-text">
-  <ng-container *nzStringTemplateOutlet="nzText">{{ nzText }}</ng-container>
-</span>
